Adds more error handling

+++ Source/Public/Remove-BuildSecrets.ps1
@@ -16,25 +16,58 @@ function Remove-BuildSecrets {
         [String]$SubscriptionID
     )
 
+    # Check if we are logged in
+    $Results = Invoke-Azcli -Arguments "account list"
+
+    if ($Results.Count -lt 1) {
+        throw "You must login first"    
+    }
+
     # Select the appropriate subscription
     if ($SubscriptionID) {
         Invoke-Azcli -Arguments "account set -s $SubscriptionID"
+    }
+
+    $Results = Invoke-Azcli -Arguments "account show"
+
+    if ($Results.state -ne 'Enabled') {
+        throw "You must select a subscription"    
     }
 
     # Get all secrets from specified vault's
     $Secrets = @()
     
     foreach ($Name in $KeyVaultName) { 
-       
-        $Secrets = Invoke-Azcli -Arguments "keyvault secret list --vault-name $Name" | ForEach-Object { Split-Path $_.id -Leaf }  
-   
-        foreach ($Secret in $Secrets) { 
 
+        $Results = Invoke-Azcli -Arguments "keyvault show --name $Name"
+        
+        if ($Results.name -ne $Name) {
+            throw "Key vault [$name] does not exists."
+        }
+
+        Write-Verbose "Removing Secrets from Vault [$Name]"       
+
+        $Results = Invoke-Azcli -Arguments "keyvault secret list --vault-name $Name"
+
+        if ($Results.Count -lt 1) {
+            Write-Verbose "No secrets found in vault [$Name]"
+        }
+
+        $Results = Invoke-Azcli -Arguments "keyvault secret list --vault-name $Name"
+        
+        $Secrets = @()
+
+        foreach ($Result in $Results) {
+            $Secrets += Split-Path $Result.id -Leaf
+        }       
+        
+        foreach ($Secret in $Secrets) {  
             $var = Get-Item -Path Env:$Secret -ErrorAction SilentlyContinue
             
             if ($var) {
                 # Set Environment Variable
                 Remove-Item -Path Env:$Secret
+                Write-Verbose "Getting secret [$Secret]"
             } else {
                 Write-Output "Could not find secret [$Secret] in current environment"
             }
